Update main.py and api_module, updated requirements and began work on documentation
To begin I have massively changed the api_module, adding the webbrowser and webview imports to improve overall use, I have added a function to open the APOD in a GUI and a function to open the link directly. In main.py I have imported the api as well as tkinter and began work on creating a simple GUI that also support previous dates.