Continued work on potentially saving the image
I have continued my work on potentially creating a way for the images to be directly saved. The current issue with urllib is that it's saving it as a tuple, apparently the image is still stored within the disk itself. I will continue working on possibly converting it into a proper image file.